transformed directory structure for frontend/backend development
