+++ frontend/generate-react-cli.json
